% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrasteps-package.R
\docType{package}
\name{extrasteps-package}
\alias{extrasteps}
\alias{extrasteps-package}
\title{extrasteps: More Miscellaneous Steps for the 'recipes' Package}
\description{
Contains additional miscellaneous steps for the 'recipes' package. These steps are useful, but doesn't have a good home in other 'recipes' packages or its extensions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EmilHvitfeldt/extrasteps}
  \item \url{https://emilhvitfeldt.github.io/extrasteps/}
  \item Report bugs at \url{https://github.com/EmilHvitfeldt/extrasteps/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

}
\keyword{internal}
