% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_stringdist.R, R/date_after.R,
%   R/date_before.R, R/date_nearest.R, R/difftime.R, R/encoding_binary.R,
%   R/encoding_frequency.R, R/maxabs.R, R/minmax.R, R/required_pkgs.R,
%   R/robust.R, R/time_event.R, R/unit_normalize.R, R/yeet.R
\name{required_pkgs.step_combine_stringdist}
\alias{required_pkgs.step_combine_stringdist}
\alias{required_pkgs.step_date_after}
\alias{required_pkgs.step_date_before}
\alias{required_pkgs.step_date_nearest}
\alias{required_pkgs.step_difftime}
\alias{required_pkgs.step_encoding_binary}
\alias{required_pkgs.step_encoding_frequency}
\alias{required_pkgs.step_maxabs}
\alias{required_pkgs.step_minmax}
\alias{required_pkgs.extrasteps}
\alias{required_pkgs.step_robust}
\alias{required_pkgs.step_time_event}
\alias{required_pkgs.step_unit_normalize}
\alias{required_pkgs.step_yeet}
\title{S3 methods for tracking which additional packages are needed for steps.}
\usage{
\method{required_pkgs}{step_combine_stringdist}(x, ...)

\method{required_pkgs}{step_date_after}(x, ...)

\method{required_pkgs}{step_date_before}(x, ...)

\method{required_pkgs}{step_date_nearest}(x, ...)

\method{required_pkgs}{step_difftime}(x, ...)

\method{required_pkgs}{step_encoding_binary}(x, ...)

\method{required_pkgs}{step_encoding_frequency}(x, ...)

\method{required_pkgs}{step_maxabs}(x, ...)

\method{required_pkgs}{step_minmax}(x, ...)

\method{required_pkgs}{step_robust}(x, ...)

\method{required_pkgs}{step_time_event}(x, ...)

\method{required_pkgs}{step_unit_normalize}(x, ...)

\method{required_pkgs}{step_yeet}(x, ...)
}
\arguments{
\item{x}{A recipe step}
}
\value{
A character vector
}
\description{
Recipe-adjacent packages always list themselves as a required package so that
the steps can function properly within parallel processing schemes.
}
\keyword{internal}
