% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_stringdist.R, R/date_after.R,
%   R/date_before.R, R/date_nearest.R, R/difftime.R, R/tidy.R, R/time_event.R,
%   R/yeet.R
\name{tidy.step_combine_stringdist}
\alias{tidy.step_combine_stringdist}
\alias{tidy.step_date_after}
\alias{tidy.step_date_before}
\alias{tidy.step_date_nearest}
\alias{tidy.step_difftime}
\alias{tidy.recipe}
\alias{tidy.step_time_event}
\alias{tidy.step_yeet}
\title{Tidy the Result of a Recipe}
\usage{
\method{tidy}{step_combine_stringdist}(x, ...)

\method{tidy}{step_date_after}(x, ...)

\method{tidy}{step_date_before}(x, ...)

\method{tidy}{step_date_nearest}(x, ...)

\method{tidy}{step_difftime}(x, ...)

\method{tidy}{step_time_event}(x, ...)

\method{tidy}{step_yeet}(x, ...)
}
\arguments{
\item{x}{A \code{step_yeet} object.}

\item{...}{Not currently used.}
}
\description{
\code{tidy} will return a data frame that contains information
regarding a recipe or operation within the recipe (when a \code{tidy}
method for the operation exists). See \link[recipes:tidy.recipe]{recipes::tidy.recipe} for
more information.
}
